+++ intelmq/bots/collectors/mail/collector_mail_attach.py
@@ -29,7 +29,7 @@ class MailAttachCollectorBot(CollectorBot):
 
                 if (self.parameters.subject_regex and
                         not re.search(self.parameters.subject_regex,
-                                      message.subject)):
+                                      re.sub("\r\n\s", " ", message.subject))):
                     continue
 
                 for attach in message.attachments: