+++ intelmq/bots/collectors/rt/collector_rt.py
@@ -80,9 +80,9 @@ class RTCollectorBot(CollectorBot):
                 raw = resp.text
 
             report = self.new_report()
-            report.add("raw", raw, sanitize=True)
-            report.add("rtir_id", ticket_id, sanitize=True)
-            report.add("time.observation", created + ' UTC', force=True)
+            report.add("raw", raw)
+            report.add("rtir_id", ticket_id)
+            report.add("time.observation", created + ' UTC', overwrite=True)
             self.send_message(report)
 
             if self.parameters.take_ticket: