+++ intelmq/bots/experts/abusix/expert.py
@@ -30,11 +30,11 @@ class AbusixExpertBot(Bot):
             abuse_contact_key = key + "abuse_contact"
             if abuse_contact_key in event and not self.parameters.overwrite:
                 continue
-            if event.contains(ip_key):
+            if ip_key in event:
                 ip = event.get(ip_key)
                 email = self.lookup(ip)
                 if email:
-                    event.add(abuse_contact_key, email, force=True)
+                    event.add(abuse_contact_key, email, overwrite=True)
 
         self.send_message(event)
         self.acknowledge_message()