+++ intelmq/bots/parsers/hphosts/parser.py
@@ -2,7 +2,6 @@
 
 from intelmq.lib import utils
 from intelmq.lib.bot import Bot
-from intelmq.lib.harmonization import IPAddress
 
 
 class HpHostsParserBot(Bot):
@@ -31,9 +30,7 @@ class HpHostsParserBot(Bot):
 
             event = self.new_event(report)
 
-            if IPAddress.is_valid(values[1]):
-                event.add("source.ip", values[1])
-            else:
+            if not event.add("source.ip", values[1], raise_failure=False):
                 event.add("source.fqdn", values[1])
 
             event.add('classification.type', 'blacklist')