+++ intelmq/lib/bot.py
@@ -379,10 +379,7 @@ class Bot(object):
 
         for option, value in config.items():
             setattr(self.parameters, option, value)
-            self.__log_buffer.append(('debug',
-                                      "Defaults configuration: parameter {!r} "
-                                      "loaded  with value {!r}.".format(option,
-                                                                        value)))
+            self.__log_configuration_parameter("defaults", option, value)
 
     def __load_system_configuration(self):
         if os.path.exists(SYSTEM_CONF_FILE):
@@ -395,9 +392,7 @@ class Bot(object):
 
             for option, value in config.items():
                 setattr(self.parameters, option, value)
-                self.__log_buffer.append(('debug',
-                                          "System configuration: parameter {!r} "
-                                          "loaded  with value {!r}.".format(option, value)))
+                self.__log_configuration_parameter("system", option, value)
 
     def __load_runtime_configuration(self):
         self.logger.debug("Loading runtime configuration from %r." % RUNTIME_CONF_FILE)
@@ -411,8 +406,7 @@ class Bot(object):
                 self.logger.warning('Old runtime configuration format found.')
             for option, value in params.items():
                 setattr(self.parameters, option, value)
-                self.logger.debug("Runtime configuration: parameter {!r} "
-                                  "loaded with value {!r}.".format(option, value))
+                self.__log_configuration_parameter("runtime", option, value)
 
     def __load_pipeline_configuration(self):
         self.logger.debug("Loading pipeline configuration from %r." % PIPELINE_CONF_FILE)
@@ -435,6 +429,18 @@ class Bot(object):
             raise ValueError("Pipeline configuration: no key "
                              "{!r}.".format(self.__bot_id))
 
+    def __log_configuration_parameter(self, config_name, option, value):
+        if "password" in option:
+            value = "HIDDEN"
+
+        message = "{} configuration: parameter {!r} loaded with value {!r}."\
+            .format(config_name.title(), option, value)
+
+        if self.logger:
+            self.logger.debug(message)
+        else:
+            self.__log_buffer.append(("debug", message))
+
     def __load_harmonization_configuration(self):
         self.logger.debug("Loading Harmonization configuration from %r." % HARMONIZATION_CONF_FILE)
         self.harmonization = utils.load_configuration(HARMONIZATION_CONF_FILE)
@@ -451,6 +457,8 @@ class Bot(object):
 
 
 class ParserBot(Bot):
+    csv_params = {}
+    ignore_lines_starting = []
 
     def __init__(self, bot_id):
         super(ParserBot, self).__init__(bot_id=bot_id)
@@ -463,7 +471,12 @@ class ParserBot(Bot):
         """
         A basic CSV parser.
         """
-        raw_report = utils.base64_decode(report.get("raw"))
+        raw_report = utils.base64_decode(report.get("raw")).strip()
+        if self.ignore_lines_starting:
+            raw_report = '\n'.join([line for line in raw_report.splitlines()
+                                    if not any([line.startswith(prefix) for prefix
+                                                in self.ignore_lines_starting])])
+
         for line in csv.reader(io.StringIO(raw_report)):
             yield line
 
@@ -471,7 +484,12 @@ class ParserBot(Bot):
         """
         A basic CSV Dictionary parser.
         """
-        raw_report = utils.base64_decode(report.get("raw"))
+        raw_report = utils.base64_decode(report.get("raw")).strip()
+        if self.ignore_lines_starting:
+            raw_report = '\n'.join([line for line in raw_report.splitlines()
+                                    if not any([line.startswith(prefix) for prefix
+                                                in self.ignore_lines_starting])])
+
         for line in csv.DictReader(io.StringIO(raw_report)):
             yield line
 
@@ -487,7 +505,9 @@ class ParserBot(Bot):
             parse = ParserBot.parse_csv
         """
         for line in utils.base64_decode(report.get("raw")).splitlines():
-            yield line.strip()
+            line = line.strip()
+            if not any([line.startswith(prefix) for prefix in self.ignore_lines_starting]):
+                yield line
 
     def parse_line(self, line, report):
         """
@@ -541,8 +561,6 @@ class ParserBot(Bot):
         writer = csv.writer(out)
         writer.writerow(line)
         return out.getvalue()
-
-    csv_params = {}
 
     def recover_line_csv_dict(self, line):
         """