+++ CHANGELOG.md
@@ -5,7 +5,7 @@ v1.0 (in development, master branch)
 ----
 
 ### General changes
-- Dropped support for Python 2
+- Dropped support for Python 2, Python >= 3.3 is needed
 - Dropped startup.conf and system.conf. Sections in BOTS can be copied directly yo runtime.conf now.
 
 ### Bot changes
@@ -29,6 +29,7 @@ v1.0 (in development, master branch)
 - STY: code obeys PEP8, new code should always be properly formatted
 - ENH: More code is Python 3 compatible
 - DOC: Updated user and dev guide
+- Removed Message.contains, Message.update methods Message.add ignore parameter
 
 ###Configuration
 - ENH: New parameter and field named accuracy to represent the accuracy of each feed